Fixes clonning records not working
